# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/analytics/process-logs": {
      "post": {
        "description": "Schedules processing of user logs for analytics",
        "operationId": "AnalyticsController_processLogs",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessLogsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Log processing scheduled"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Process user logs",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/job/{id}": {
      "get": {
        "description": "Gets the status of a scheduled job",
        "operationId": "AnalyticsController_getJobStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status"
          },
          "404": {
            "description": "Job not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get job status",
        "tags": [
          "analytics"
        ]
      }
    },
    "/webhooks/clerk": {
      "post": {
        "description": "Processes webhooks from Clerk for user and organization events",
        "operationId": "ClerkWebhooksController_handleWebhook",
        "parameters": [
          {
            "name": "svix-signature",
            "required": true,
            "in": "header",
            "description": "Svix signature for webhook verification",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "svix-id",
            "required": true,
            "in": "header",
            "description": "Svix ID for webhook verification",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "svix-timestamp",
            "required": true,
            "in": "header",
            "description": "Svix timestamp for webhook verification",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Clerk webhook payload",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully"
          },
          "400": {
            "description": "Bad request - missing headers"
          },
          "401": {
            "description": "Unauthorized - invalid signature"
          }
        },
        "summary": "Handle Clerk webhook",
        "tags": [
          "webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wally API",
    "description": "The API documentation for the Wally application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
      "ProcessLogsDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID to process logs for",
            "example": "user_123456789"
          },
          "timeframe": {
            "type": "string",
            "description": "Timeframe for log processing",
            "enum": [
              "day",
              "week",
              "month"
            ],
            "default": "day",
            "example": "day"
          }
        },
        "required": [
          "userId"
        ]
      },
...
    }
  }
}